uspace: fix compiler warnings toward a zero-warning build
Resolve every -Wall -Wextra warning on both gcc and clang (xdp and
no-xdp) so -Werror can be turned on in CI. Fixes only; the two classes
that stay suppressed are handled separately in the CI config.

Notable real bug: ec_eoe_init() passed a computed interface name as the
printf format string (snprintf(..., name)); now "%s", name.

- stringop-truncation: NUL-safe copies. tool_api.c/master.c use
  bound = size-1 with an explicit terminator; pal_eoe.c/cdev.c/
  transport_xdp.c switch to snprintf / bounded memcpy (the
  "may-be-truncated" variant an explicit terminator cannot silence).
- sign-compare: size/offset comparisons get explicit (size_t) casts
  (fsm_slave_scan.c, fsm_soe.c, fsm_master.c, fsm_foe.c, ethernet.c,
  slave.c); master.c uses (ec_time_t) 1 so both max() operands match.
- clang-only: fsm_slave_config.c passes NULL (not EC_DIR_INVALID) for
  the sync_config pointer, matching the sibling std-mailbox calls;
  domain.c drops the set-but-unused datagram_count.
- missing-field-initializers: fully initialize the ec_sync_info_t
  {0xff} terminators (fsm_change.c code table and the test fixtures).

All four CI build combos are warning-free; 15/15 unit tests pass; the
kernel module build still compiles the shared master.c change.

Refs #182

+++ master/fsm_slave_scan.c
@@ -671,13 +671,13 @@ void ec_fsm_slave_scan_state_sii_data(ec_fsm_slave_scan_t *fsm
 
     // 2 words fetched
 
-    if (fsm->sii_offset + 2 <= slave->sii_nwords) { // 2 words fit
+    if ((size_t) fsm->sii_offset + 2 <= slave->sii_nwords) { // 2 words fit
         memcpy(slave->sii_words + fsm->sii_offset, fsm->fsm_sii.value, 4);
     } else { // copy the last word
         memcpy(slave->sii_words + fsm->sii_offset, fsm->fsm_sii.value, 2);
     }
 
-    if (fsm->sii_offset + 2 < slave->sii_nwords) {
+    if ((size_t) fsm->sii_offset + 2 < slave->sii_nwords) {
         // fetch the next 2 words
         fsm->sii_offset += 2;
         ec_fsm_sii_read(&fsm->fsm_sii, slave, fsm->sii_offset,
@@ -786,7 +786,7 @@ void ec_fsm_slave_scan_state_sii_data(ec_fsm_slave_scan_t *fsm
     while (EC_READ_U16(cat_word) != 0xFFFF) {
 
         // type and size words must fit
-        if (cat_word + 2 - slave->sii_words > slave->sii_nwords) {
+        if ((size_t) (cat_word + 2 - slave->sii_words) > slave->sii_nwords) {
             EC_SLAVE_ERR(slave, "Unexpected end of SII data:"
                     " Category header incomplete.\n");
             goto end;
@@ -796,7 +796,7 @@ void ec_fsm_slave_scan_state_sii_data(ec_fsm_slave_scan_t *fsm
         cat_size = EC_READ_U16(cat_word + 1);
         cat_word += 2;
 
-        if (cat_word + cat_size - slave->sii_words > slave->sii_nwords) {
+        if ((size_t) (cat_word + cat_size - slave->sii_words) > slave->sii_nwords) {
             EC_SLAVE_WARN(slave, "Unexpected end of SII data:"
                     " Category data incomplete.\n");
             goto end;
@@ -836,7 +836,7 @@ void ec_fsm_slave_scan_state_sii_data(ec_fsm_slave_scan_t *fsm
         }
 
         cat_word += cat_size;
-        if (cat_word - slave->sii_words >= slave->sii_nwords) {
+        if ((size_t) (cat_word - slave->sii_words) >= slave->sii_nwords) {
             EC_SLAVE_WARN(slave, "Unexpected end of SII data:"
                     " Next category header missing.\n");
             goto end;
